pull out some functions to task_handling.py, change new_competitor to add if new
